add front-js

+++ src/main/java/org/corsojava/pizzeria/api/controller/ApiPizzeriaController.java
@@ -3,7 +3,9 @@ package org.corsojava.pizzeria.api.controller;
 import java.util.List;
 import java.util.Optional;
 
+import org.corsojava.pizzeria.model.Ingredient;
 import org.corsojava.pizzeria.model.Pizzeria;
+import org.corsojava.pizzeria.repository.IngredientRepository;
 import org.corsojava.pizzeria.repository.PizzeriaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -25,6 +27,9 @@ public class ApiPizzeriaController {
 	
 	@Autowired
 	PizzeriaRepository pizzeriarep;
+	
+	@Autowired
+	IngredientRepository ingredientrep;
 	
 	@GetMapping() 
 	public List<Pizzeria> index(){
@@ -55,10 +60,23 @@ public class ApiPizzeriaController {
 		return pizzeriarep.save(p);
 	}
 	
-	@DeleteMapping("{id}")       // Elimina pizza
-	public void delete(
-			@PathVariable("id") Integer id) {
-		pizzeriarep.deleteById(id);
+//	@DeleteMapping("/delete/{id}")       // Elimina pizza
+//	public void delete(
+//			@PathVariable("id") Integer id) {
+//		pizzeriarep.deleteById(id);
+//	}
+	
+	@DeleteMapping("/delete/{id}")	//eliminazione   Richiesta DELETE /api/xx
+	public ResponseEntity<Pizzeria> delete(
+			@PathVariable("id") Integer id) {		
+		Optional<Pizzeria> res =pizzeriarep.findById(id);		
+		if (res.isPresent()) {			
+			pizzeriarep.deleteById(id);
+			return new ResponseEntity<Pizzeria>( HttpStatus.OK);
+		}			
+	    else
+	    	return new ResponseEntity<Pizzeria>(HttpStatus.NOT_FOUND);
+		
 	}
